SMT 01.17.15 - Solve bug and add alert on remove seed feature

+++ src/views/Seeds.vue
@@ -83,11 +83,13 @@ export default {
 					}
 				})
 			}).catch(error => { this.error = error })
-		}
+		},
 		remove (id) {
-			axios.delete('product/' + id + '?_format=json').then(() => {
-				this.list()
-			}).catch(error => { this.error = error })	
+			if (confirm("Tem certeza que deseja excluír?")) {
+				axios.delete('product/' + id + '?_format=json').then(() => {
+					this.list()
+				}).catch(error => { this.error = error })	
+			}
 		}
 	}
